Break out confest into pytest CLI options and fixtures/ directory
